Bump `sgx`, `sallyport` and `noted` deps
This requires bumps of a few other crates too. All input parameters from
the shims are now covered by the versioned sallyport.

+++ internal/shim-sgx/src/handler/base.rs
@@ -62,7 +62,7 @@ impl<'a> BaseSyscallHandler for super::Handler<'a> {
         debug!(self, "{}(", name);
         for (i, arg) in argv[..argc].iter().copied().enumerate() {
             let prefix = if i > 0 { ", " } else { "" };
-            debug!(self, "{}0x{:x}", prefix, u64::from(arg));
+            debug!(self, "{}0x{:x}", prefix, arg);
         }
 
         debugln!(self, ")");
